Improve app_block runner helper modules
