+++ vemonitor_m8/core/app_run_main.py
@@ -1,0 +1,158 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+"""App block run Helper"""
+import logging
+from typing import Optional
+from vemonitor_m8.core.utils import Utils as Ut
+from vemonitor_m8.events.app_block_events import AppBlockEvents
+from vemonitor_m8.core.data_cache import DataCache
+from vemonitor_m8.workers.redis.redis_cache import RedisCache
+from vemonitor_m8.models.inputs_cache import InputsCache
+from vemonitor_m8.core.data_checker import DataChecker
+from vemonitor_m8.models.config import Config
+from vemonitor_m8.workers.workers_manager import WorkersManager
+from vemonitor_m8.core.exceptions import DeviceInputValueError
+from vemonitor_m8.core.exceptions import RedisConnectionException
+from vemonitor_m8.core.exceptions import SettingInvalidException
+from vemonitor_m8.core.exceptions import DeviceDataConfError
+
+__author__ = "Eli Serra"
+__copyright__ = "Copyright 2022, Eli Serra"
+
+logging.basicConfig()
+logger = logging.getLogger("vemonitor")
+
+
+class AppBlockRun:
+    """Async App block run Helper"""
+
+    def __init__(self, conf: Config):
+        self._run = False
+        self.conf: Optional[Config] = None
+        self.events = AppBlockEvents()
+        self.inputs_data: Optional[InputsCache] = None
+        self.workers = WorkersManager()
+        if self.set_conf(conf)\
+                and self.init_data_cache():
+            self._run = True
+
+    def is_ready(self) -> bool:
+        """Test if class instance have valid battery_bank property."""
+        return self.is_conf_ready()\
+            and self.is_cache_ready()\
+            and self._run is True
+
+    def is_conf_ready(self) -> bool:
+        """Test if class instance have valid battery_bank property."""
+        return AppBlockRun.is_conf(self.conf)
+
+    def is_cache_ready(self) -> bool:
+        """Test if class instance have valid battery_bank property."""
+        return isinstance(self.inputs_data, InputsCache)
+
+    def init_redis_cache(self) -> bool:
+        """Init redis cache object."""
+        result = False
+        if self.is_conf_ready():
+            redis_cache = self.conf.get_redis_cache_by_key(  # type: ignore
+                index=0
+            )
+            if Ut.is_dict(redis_cache, not_null=True):
+                try:
+                    self.inputs_data = RedisCache(
+                        max_rows=redis_cache.get("max_data_points"),
+                        connector=self.get_app_connector_by_key_item(
+                            item_key="redis",
+                            source=redis_cache.get("source")
+                        )
+                    )
+                    result = True
+                    logger.info(
+                        "Redis Cache is enabled and active"
+                    )
+                except RedisConnectionException as ex:
+                    result = False
+                    logger.error(
+                        "Error Redis Cache is enabled, "
+                        "but we are unable to connect to Redis server."
+                        "Please check Redis App Connectors Configuration."
+                        "Exception: {%s}",
+                        ex
+                    )
+        return result
+
+    def init_data_cache(self) -> bool:
+        """
+        Init dataCache object.
+        ToDo: Redis cache is only needed for worker outputs who send bulk data
+        """
+        result = False
+        if self.is_conf_ready():
+            if self.init_redis_cache() is True:
+                result = True
+            else:
+                logger.info(
+                    "Start Memory Data Cache..."
+                )
+                self.inputs_data = DataCache(max_rows=15)
+                result = True
+        return result
+
+    def set_conf(self, conf) -> bool:
+        """Set Configuration data."""
+        result = False
+        if AppBlockRun.is_conf(conf):
+            self.conf = conf
+            result = True
+        else:
+            raise SettingInvalidException(
+                "[veMonitor] Fatal error: "
+                "Config is not valid."
+            )
+        return result
+
+    def format_input_data(self,
+                          data: dict,
+                          columns: Optional[list] = None
+                          ) -> dict:
+        """Format input data."""
+        result = None
+        data = Ut.get_items_from_dict(data, columns)
+        if Ut.is_dict(data, not_null=True):
+            try:
+                result = DataChecker.check_input_columns(
+                    data,
+                    self.conf.data_structures.get('points')
+                )
+            except (
+                DeviceDataConfError,
+                DeviceInputValueError
+            ) as ex:
+                raise DeviceDataConfError(
+                    "Fatal Error: Device Data Error. "
+                    "See Your device data configuration. "
+                    "Or some ipnut value is bad type"
+                    f"data checked: {data}"
+                ) from ex
+
+        return result
+
+    def get_app_connector_by_key_item(self, item_key: str, source: str):
+        """Get connector by key item and source."""
+        return self.conf.get_app_connector_by_key_item(
+            item_key,
+            source
+        )
+
+    def is_worker_data_ready(self):
+        """On worker data ready event"""
+        if Ut.is_dict(self.inputs_data.data, not_null=True)\
+                and len(self.inputs_data.data) >= 5:
+            self.events.on_worker_data_ready()
+
+    @staticmethod
+    def is_conf(conf: Optional[Config]) -> bool:
+        """Test if valid conf"""
+        return isinstance(conf, Config)\
+            and conf.is_valid()\
+            and len(conf.app_blocks) == 1  # type: ignore
